Stop calling the D3 force worker, which we no longer use. (#46)
We were already fixing every single Node in place with the `fx` and `fy`
parameters for the D3 force thingy. This meant that we were calling into
it, and it was in turn doing nothing for some number of iterations.
We don't really need to be doing that.

This change directly uses the `[x, y]` positions we get from the server
to position Nodes and eliminates all use of the d3 force worker thing.